#50 Address Copilot round 11 review feedback
MenuCrawler:
- Remove Finder-specific cache duration (use unified 0.5s)
- Rely on explicit cache invalidation in ViewModel for Finder

MockMenuItemFactory:
- Add "[Mock] " prefix to item titles for clear identification
- Update path to "Mock Menu" for consistency

HotkeyManager:
- Use Int64 consistently for keyCode comparison

PanelController:
- Use explicit allowedSystemModifiers (.numericPad, .function, .capsLock)
- Subtract allowed modifiers instead of checking intersection

ScrollBehaviorUITests:
- Add mockItemsLoadTimeout constant (2.0s)

+++ Portal/Services/HotkeyManager.swift
@@ -169,7 +169,8 @@ final class HotkeyManager {
 
     private func isHotkeyEvent(_ event: NSEvent) -> Bool {
         let modifiers = event.modifierFlags.intersection(.deviceIndependentFlagsMask)
-        return modifiers == .option && event.keyCode == UInt16(kSpaceKeyCode)
+        // Use Int64 for consistency with CGEventTap callback comparison
+        return modifiers == .option && Int64(event.keyCode) == kSpaceKeyCode
     }
 
     deinit {
